Add files via upload
Overview
To create a personalised chatbox for new University students. To aid new students, who have just started a new chapter in their lives. First days at the University are full of excitement, but, equally, it could feel rather daunting and overwhelming for freshers.
UniBuddy offers a personalised greeting to a newcomer and it is designed to make the transition smoother for freshmen.